13ai/requirements.txt
numpy==1.26.4
opencv-python-headless==4.10.0.84
Pillow==10.4.0
ultralytics==8.3.4
websockets==13.0.1
python-socketio==5.11.4
eventlet==0.36.0